[pipelines] Set up a pleroma pipeline
In preparation for moving other jobs out of the barrucadu.co.uk and
lainon.life pipelines.

Unfortunately I can't use the `library.deploy_docker_systemd_job`
here, because it assumes that the name is `deploy-{image}` and making
that not the case would have complicated it... I definitely didn't
have future extensibility in mind when I wrote many of these library
functions.